+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - SparseTensor list operators (cat, mean, sum, var)
 - MinkowskiStack containers
 - Replace all at::cuda::getCurrentCUDASparseHandle with custom getCurrentCUDASparseHandle (issue #308)
+- fix coordinate manager kernel map python function
 
 ## [0.5.1]